Remove acctype from translate_address
Threading through the AccType all the way from the decode function to
the memory helpers is a bit awkward in the proof.  In the actual address
translation, the AccType shouldn't affect the value of the returned
physical address (maybe with the exception of AccType_UNPRIV), only
whether a fault occurs or not.

Also adapt to a change in the abstract model to thread through the
virtual address to access_enabled to make clear which virtual address
was used for the bounds check.