Add assumption that UNKNOWN capabilities are derivable
Some instruction specifications allow writing UNKNOWN capabilities, e.g.
if an attempt is made to read from and write to a register at the same
time.  The specification document has a rule saying that this "must not
increase capability rights available to software";  the derivability
assumption is one way of formalising this rule.