Make use of load_cap.patch in arm-morello-dropzone
The ASL was changed recently to check the LOAD_CAP permissions after the
load, and clear the tag if it is missing.  The T-CHERI properties
currently require the permission to be checked before a tagged load,
however.  We probably want to tweak the properties to support
tag-clearing after loading, but to make progress with the proof, the
patched ASL only requests a tagged load if the permission is present,
and a data-only load otherwise.