Fix AArch64.TakeException proof
Add lemmas for the exception case of writing PCC, and add a proof and a
register accessibility assumption to the AArch64.TakeException lemma.
The latter leads to additional register accessibility conditions for
other lemmas.  This patch adds those for the manually patched functions.
For functions and instructions with loops, a register accessibility loop
invariant may now be required; it should be possible to handle this
automatically in the traces_enabledI tactic.