Prove the simplest case of fast representation
There are three cases recognised by CapIsRepresentableFast, one
where the exponent is high, one where the increment is small and
positive, and one where it is small and negative. The high proof
unfortunately needs to be divided into cases for the different
exponents (48, 49, 50) and of these the 48 case is fairly
involved.