Initial work on page table walk determinism
The plan is to show that the page table walk is deterministic,
so we can use the translated definition as the definition of
what is mapped. The plan is to extend the monad_return_rel
mechanism previously applied to CapGetBounds, with additions for
register and memory fetches (which will need to be deterministic)
and for exceptions (which are caught and handled within the walk
function). Work in progress.