Tweak system register write proof
Writing to RMR_ELx can cause a warm reset of the system, which is
non-monotonic in a way that is not captured by our properties.  Such a
write does require system register access permission, though.  Tweak the
property of the system register write instruction to show monotonicity
if either the write is to a register other than RMR_ELx, or PCC does not
have system access permission (so that a write to RMR_ELx will cause an
exception rather than a reset).