Clarify some assumptions
In particular, tag setting instructions check both system access and tag
setting is enabled.  We assume that either the one or the other is
disabled.  Note that there don't seem to be control register bits to
disable tag setting in EL3, so we assume that we are not in EL3 if
system access is enabled.