Elaborate assumption about UNKNOWN capabilities
The (prose) spec is very loose about UNKNOWN capabilities; it basically
allows them to take any value, as long as these don't increase the
capability rights available to software.  In particular, it doesn't
explicitly require these capabilities to come from input registers of
the instruction using an UNKNOWN capability.

It would be desirable to constrain this more tightly and explicitly in
future architecture versions.

In the meantime, we model this by assuming that UNKNOWN capabilities are
drawn from an arbitrary but fixed set of capabilities, and the
monotonicity theorem still holds with the additional assumption that
this set of UNKNOWN capabilities is a subset of the initially reachable
capabilities.