Progress on CapSetBounds using brute force SMT
By case-splitting on the result of count-leading-zeros and thus
the computed exponent, split into 100+ goals each of which can be
solved by SMT oracle, if one tweaks their CVC4 install a little.
It's possible these goals can also be solved by circuitisation and
bit-blasting, if one had a lot of CPU time available.

This is a partial solution to the encode/decode aspect of the
CapSetBounds issue. The proof that the resulting exponent is no
greater than the original one remains.