Start on set-bounds exponent near-unique prop
This relates to an idea for how to show the 'optimality'
property for the exponent in CapSetBounds. The process of
aligning and down-shifting can be done one bit at a time,
which clarifies what's going on. This provides the first
proofs about this process, but in hindsight they're out
by a special factor of 3.