Better approach to set-bounds exponent
With this proven, CHERI_Instantiation is now sorry-free. It still
depends on the use of the SMT oracle to prove a handful of facts
about CapSetBounds.

The better approach is to think about "sufficiency" of an exponent
to encode a particular upper/lower bound, which is a more obviously
monotonic property. The key fact is that CapSetBounds ensures that
the chosen exponent is sufficient and that the next smallest
exponent is insufficient.