Better way of hacking CVC4 version
This does the hacking of which version is picked in ML code
inside Isabelle. It still needs a cvc4-1.8 present in the
Isabelle contrib dir. It will end up there by default if the
user has the contrib in their ~/.isabelle ($ISABELLE_HOME_USER)
and has both Isabelle2021 and Isabelle2020 installed. There are
various other ways to hack it.