CHERI_Instantiation passes
The good news is, New_CVC4 isn't needed any more, and although
it was a bit painful, we got there in the end.

The bad news is, a couple of goals that were previously
bit-blastable are now done with smt_word instead. This is
probably regression in the bit-blasting, but it's tricky to
debug. Also, the "brute force" proof seems to have gotten
2-3x slower, probably because of changes to default simp rules.