Fix a proof
asl_to_sail has recently become smarter when generating assertions to
make the model type-check, and this lead to a redundant assertion
(exponents are non-negative and bounded by a maximum value) that this
proof made use of disappearing from the body of the function.  A lemma
corresponding to the missing assertion exists, and supplying it at the
right point fixes the proof.