Handle untagged memory capabilities in invocation case split property
The instantiation of the sets of invoked capabilities only talks about
tagged capabilities, but in the cases of indirect sentries there might
be invocations involving some tagged and some untagged capabilities
loaded from memory.

In the Points-to-Pair case, either the code or data capability might be
untagged in memory.

In the Points-to-PCC case, if the indirect sentry does not have the
`LOAD_CAP` permission, the code capability loaded from memory will have
its tag cleared, but the tagged indirect sentry itself will still be
unsealed and written to IDC.

Hence, make the characterisation of the sets of invoked code and data
capbilities in those cases of the property depend on that permission bit
as well the original in-memory tags, and tweak the position of the check
for the permission bit in the original instantiation of the sets of
invoked capabilities.