Tweak handling of virtual addresses for indirect sentry invocations
Require specific address for code caps in load pair instructions, even
in the case of not invoking a direct sentry, but where the loaded cap
happens to be a direct sentry.

Moreover, in the invocation case split lemma, be more explicit about the
physical addresses.  The set of accessed physical addresses will always
have at most one or two elements, whereas for virtual addresses, that
would be tricky to state with potentially non-injective address
mappings.